Refactor voxelising to loosen coupling
