+++ src/voxel_mesh.ts
@@ -2,12 +2,10 @@ import { RenderBuffer, AttributeData } from './buffer';
 import { AppConfig } from './config';
 import { GeometryTemplates } from './geometry';
 import { HashMap } from './hash_map';
-import { MaterialType, Mesh, SolidMaterial, TexturedMaterial } from './mesh';
+import { Mesh } from './mesh';
 import { OcclusionManager } from './occlusion';
-import { Axes, generateRays, rayIntersectTriangle } from './ray';
-import { Texture, TextureFiltering } from './texture';
-import { Triangle, UVTriangle } from './triangle';
-import { Bounds, LOG, RGB, UV } from './util';
+import { TextureFiltering } from './texture';
+import { Bounds, RGB } from './util';
 import { Vector3 } from './vector';
 
 export interface Voxel {
@@ -28,30 +26,17 @@ export class VoxelMesh {
     private _voxelMeshParams: VoxelMeshParams;
 
     private _voxelSize: number;
-    private _desiredHeight: number;
     private _voxels: Voxel[];
     private _voxelsHash: HashMap<Vector3, number>;
-    private _loadedTextures: { [materialName: string]: Texture };
     private _bounds: Bounds;
 
-
-    public static createFromMesh(mesh: Mesh, voxelMeshParams: VoxelMeshParams) {
-        const voxelMesh = new VoxelMesh(mesh, voxelMeshParams);
-        voxelMesh._voxelise();
-        return voxelMesh;
-    }
-
-    private constructor(mesh: Mesh, voxelMeshParams: VoxelMeshParams) {
-        LOG('New voxel mesh');
-        
+    public constructor(mesh: Mesh, voxelMeshParams: VoxelMeshParams) {
         this._mesh = mesh;
         this._voxelMeshParams = voxelMeshParams;
 
         this._voxelSize = 8.0 / Math.round(voxelMeshParams.desiredHeight);
-        this._desiredHeight = this._voxelMeshParams.desiredHeight;
         this._voxels = [];
         this._voxelsHash = new HashMap(2048);
-        this._loadedTextures = {};
         this._bounds = Bounds.getInfiniteBounds();
     }
 
@@ -67,86 +52,7 @@ export class VoxelMesh {
         return this._mesh;
     }
 
-    private _voxelise() {
-        LOG('Voxelising');
-
-        const scale = (this._desiredHeight - 1) / Mesh.desiredHeight;
-        const offset = (this._desiredHeight % 2 === 0) ? new Vector3(0.0, 0.5, 0.0) : new Vector3(0.0, 0.0, 0.0);
-        const useMesh = this._mesh.copy();
-        for (let i = 0; i < useMesh.vertices.length; ++i) {
-            useMesh.vertices[i].mulScalar(scale).add(offset);
-        }
-
-        useMesh.tris.forEach((tri, index) => {
-            const material = useMesh.materials[tri.material];
-            if (material.type == MaterialType.textured) {
-                if (!(tri.material in this._loadedTextures)) {
-                    this._loadedTextures[tri.material] = new Texture(material.path);
-                }
-            }
-            const uvTriangle = useMesh.getUVTriangle(index);
-            this._voxeliseTri(uvTriangle, material, tri.material);
-        });
-    }
-
-    private _voxeliseTri(triangle: UVTriangle, material: (SolidMaterial | TexturedMaterial), materialName: string) {
-        const rayList = generateRays(triangle.v0, triangle.v1, triangle.v2);
-
-        rayList.forEach((ray) => {
-            const intersection = rayIntersectTriangle(ray, triangle.v0, triangle.v1, triangle.v2);
-            if (intersection) {
-                let voxelPosition: Vector3;
-                switch (ray.axis) {
-                case Axes.x:
-                    voxelPosition = new Vector3(Math.round(intersection.x), intersection.y, intersection.z);
-                    break;
-                case Axes.y:
-                    voxelPosition = new Vector3(intersection.x, Math.round(intersection.y), intersection.z);
-                    break;
-                case Axes.z:
-                    voxelPosition = new Vector3(intersection.x, intersection.y, Math.round(intersection.z));
-                    break;
-                }
-
-                let voxelColour: RGB;
-                if (this._voxelMeshParams.useMultisampleColouring && material.type === MaterialType.textured) {
-                    const samples: RGB[] = [];
-                    for (let i = 0; i < AppConfig.MULTISAMPLE_COUNT; ++i) {
-                        const samplePosition = Vector3.add(voxelPosition, Vector3.random().addScalar(-0.5));
-                        samples.push(this._getVoxelColour(triangle, material, materialName, samplePosition));
-                    }
-                    voxelColour = RGB.averageFrom(samples);
-                } else {
-                    voxelColour = this._getVoxelColour(triangle, material, materialName, voxelPosition);
-                }
-                this._addVoxel(voxelPosition, voxelColour);
-            }
-        });
-    }
-
-    private _getVoxelColour(triangle: UVTriangle, material: (SolidMaterial | TexturedMaterial), materialName: string, location: Vector3): RGB {
-        if (material.type == MaterialType.solid) {
-            return material.colour;
-        }
-
-        const area01 = new Triangle(triangle.v0, triangle.v1, location).getArea();
-        const area12 = new Triangle(triangle.v1, triangle.v2, location).getArea();
-        const area20 = new Triangle(triangle.v2, triangle.v0, location).getArea();
-        const total = area01 + area12 + area20;
-
-        const w0 = area12 / total;
-        const w1 = area20 / total;
-        const w2 = area01 / total;
-
-        const uv = new UV(
-            triangle.uv0.u * w0 + triangle.uv1.u * w1 + triangle.uv2.u * w2,
-            triangle.uv0.v * w0 + triangle.uv1.v * w1 + triangle.uv2.v * w2,
-        );
-            
-        return this._loadedTextures[materialName].getRGB(uv, this._voxelMeshParams.textureFiltering);
-    }
-
-    private _addVoxel(pos: Vector3, colour: RGB) {
+    public addVoxel(pos: Vector3, colour: RGB) {
         const voxelIndex = this._voxelsHash.get(pos);
         if (voxelIndex !== undefined) {
             const voxel = this._voxels[voxelIndex];
